Fixed sampling for env maps + exit portals so it is efficient again. Changed object surface area array to use floats instead of doubles. Cleaned up and rearranged Object a bit.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@5088 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/geometry.h
@@ -82,7 +82,7 @@ public:
 	
 	virtual unsigned int getMaterialIndexForTri(unsigned int tri_index) const { return 0; }
 
-	virtual void getSubElementSurfaceAreas(const Matrix4f& to_parent, std::vector<double>& surface_areas_out) const = 0;
+	virtual void getSubElementSurfaceAreas(const Matrix4f& to_parent, std::vector<float>& surface_areas_out) const = 0;
 	
 	// Sample the surface of the given sub-element.
 	virtual void sampleSubElement(unsigned int sub_elem_index, const SamplePair& samples, Pos3Type& pos_out, Vec3Type& normal_out, HitInfo& hitinfo_out, float sub_elem_area_ws, Real& p_out) const = 0;
